requirements.txt
Flask
Flask-Cors
requests
beautifulsoup4
google-generativeai
python-dotenv